WQP-1592 (Drop "unlogged" keyword from all WQP tables)    Removed unlogged from usCountiesDis20121015.sql    Removed unlogged from tl2019UsCountyGeopkg.sql    Updated id to version v2 for the create changeLogs
The approach taken here is that, since us_counties_dis_20121015 is
a regular table with no partitions (fairly small, 47M currently on test),
the setting of the table to logged can be done separately using a single
sql command one time for the existing databases. Going forward when this is
used to create the table, the table will be in the desired state.
No need to maintain a separate lquidbase script.